requirements.txt
undetected-chromedriver
selenium
requests
pandas
openpyxl
pyotp
python-dateutil